BUG: called wrong variable
Oops I called the wrong variable in my if statement. changed the n to
an x in the if statement contained in the counted loop so that each
time x is incremented it evaluates the number that the counter is on
instead of n which is passed via the GET request.