BUG: entities were not passing into each of the blog posts at all
created global function called render_str and changed the current
render_str to render_tmp so that later on there would not be any
confusion as to which render_str was being called. Also made the same
change to the render function under the Post class changing it to
render_post thus avoiding more confusion with the render function under
the TemplateHandler class.